requirements.txt
docker-py==1.10.2
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
SQLAlchemy==1.0.15
WTForms==2.1
Werkzeug==0.11.11
